Make more updates for SW Internal error handling

+++ Master/Components/DeviceControl/Include/Global/DeviceControlGlobal.h
@@ -463,19 +463,19 @@ const quint32 DEVICE_INSTANCE_ID_CAN_COMMUTOR   = 0x000080C5;   //!< CAN Communi
 const quint32 DEVICE_INSTANCE_ID_OTHER_DEVICE   = 0x000080C6;   //!< Other device
 
 /*! Event group */
-const quint32  EVENT_GRP_DCL_DEVCTRL            = 0x0301; //!< device control
-const quint32  EVENT_GRP_DCL_CANINTERFACE       = 0x0302;  //!< can interface
-const quint32  EVENT_GRP_DCL_CONFIGURATION      = 0x0303;  //!< configuration
-const quint32 EVENT_GRP_DCL_DEVICEPROC_CONFIG   = 0x0304; //!< device processing - configuration
-const quint32 EVENT_GRP_DCL_CANBUS              = 0x0305; //!< can bus
-const quint32 EVENT_GRP_DCL_DIAG                = 0x0306; //!< diagnsostic / service
-const quint32 EVENT_GRP_DCL_RV_DEV              = 0x0307; //!< Rotary Valve
-const quint32 EVENT_GRP_DCL_AL_DEV              = 0x0308; //!< Air liquid
-const quint32 EVENT_GRP_DCL_RT_DEV              = 0x0309; //!< Retort
-const quint32 EVENT_GRP_DCL_OVEN_DEV            = 0x030A; //!< Oven
-const quint32 EVENT_GRP_DCL_MC_DEV              = 0x030B; //!< Main controll
-const quint32 EVENT_GRP_DCL_NODE_DCL            = 0x030C; //!< CANNode
-const quint32 EVENT_GRP_DCL_NODE_SLV            = 0x030D; //!< CANNode
+const quint16  EVENT_GRP_DCL_DEVCTRL            = 0x0301; //!< device control
+const quint16  EVENT_GRP_DCL_CANINTERFACE       = 0x0302;  //!< can interface
+const quint16  EVENT_GRP_DCL_CONFIGURATION      = 0x0303;  //!< configuration
+const quint16  EVENT_GRP_DCL_DEVICEPROC_CONFIG  = 0x0304; //!< device processing - configuration
+const quint16  EVENT_GRP_DCL_CANBUS             = 0x0305; //!< can bus
+const quint16  EVENT_GRP_DCL_DIAG               = 0x0306; //!< diagnsostic / service
+const quint16  EVENT_GRP_DCL_RV_DEV             = 0x0307; //!< Rotary Valve
+const quint16  EVENT_GRP_DCL_AL_DEV             = 0x0308; //!< Air liquid
+const quint16  EVENT_GRP_DCL_RT_DEV             = 0x0309; //!< Retort
+const quint16  EVENT_GRP_DCL_OVEN_DEV           = 0x030A; //!< Oven
+const quint16  EVENT_GRP_DCL_MC_DEV             = 0x030B; //!< Main controll
+const quint16  EVENT_GRP_DCL_NODE_DCL           = 0x030C; //!< CANNode
+const quint16  EVENT_GRP_DCL_NODE_SLV           = 0x030D; //!< CANNode
 
 /*! Slaves used in Himalaya project */
 typedef enum {
